backend/requirements.txt
fastapi
uvicorn
tortoise-orm
pydantic
asyncmy
python-dotenv
faker